Fix null neededexports overriding (#13)
Currently, if needed exports is null, the preset value overrides the
user value. As such, if the user value is null (trying to export all),
the dad-preset would override it to export only the preset dirs.

This is still the case with this PR, but now,
neededexportsfilepath='All' is supported. This can be put in the user
path to export all files. Its also the new value in the
template.appsetting.