Faster PNG encode with SKPixmap (#16)
Added faster way of doing PNG encoding by using SKPixmap instead of
SKImage since we don't want an image before we encode anyway and
SKPixmap has more control options like which PNG filters are used and
the level of intensity of the compression algorithm usually called
DEFLATE. For our purposes no filters are used and the compression is set
to 1 (0 is the lowest, 9 is highest). This is a good balance between
speed and file size as setting it to 0 results in very large files.