Added parallel processing for extracting assets (#15)
Processing files is now done in parallel via a queue that contains file
paths as strings. This was the decided approach over batching as it has
much better load balancing since each new thread processes the next file
in the queue with little to no dependence on the overall file structure.
This way threads that get stuck for a while don't block other threads
that may process many small files.